+++ backend/ecommerce/src/main/java/com/example/ecommerce/repository/ProductReviewRepository.java
@@ -8,6 +8,6 @@ import com.example.ecommerce.entity.ProductReview;
 
 public interface ProductReviewRepository extends JpaRepository<ProductReview, Long> {
 
-	List<ProductReview> findByProductId(Long productId);
+	List<ProductReview> findByProductId(int productId);
     // Add any additional custom query methods here
 }
